Update transactions list component to use flat list
